Change update logic to make the rich presence more responsive when the data changes
Instead of only polling every 5 seconds, it now reads data from the game every 0.5 seconds, and asks for an update if the important data (currently gamestate, stagestate, stage, lives, bombs, and gameovers) has changed. There is still a regular update to make sure everything else has the opportunity to be updated, but it's now every 8 seconds. It should be set up to stay within a limit of 5 updates sent per 20 seconds.