Added photo counts / item uses to StateData
Before, we had to wait until the next refresh on the RPC display to see the photo counts (th095, th125) and item usage (th143) be updated. They are now stored in StateData and updated immediately.
Since we limit the number of updates in time intervals, this can also cause the update to fall behind when a lot of RPC updates are done at the same time.